rebasing

+++ test/dropout_util.hpp
@@ -76,7 +76,8 @@ inline void mat_vec(const unsigned int* matrix, unsigned int* vector)
             {
                 std::transform(result,
                                result + XORWOW_DIM,
-                               matrix + (XORWOW_DIM * (i * XORWOW_BITS + j)),
+                               matrix +
+                                   static_cast<std::ptrdiff_t>(XORWOW_DIM * (i * XORWOW_BITS + j)),
                                result,
                                std::bit_xor<unsigned int>{});
             }
@@ -89,7 +90,7 @@ inline void mat_mat(unsigned int* matrixA, const unsigned int* matrixB)
 {
     for(int i = 0; i < XORWOW_DIM * XORWOW_BITS; i++)
     {
-        mat_vec(matrixB, matrixA + i * XORWOW_DIM);
+        mat_vec(matrixB, matrixA + static_cast<std::ptrdiff_t>(i * XORWOW_DIM));
     }
 }
 
@@ -113,7 +114,9 @@ inline void mat_pow(unsigned int* matrixP, const unsigned int* matrix, unsigned 
 
     unsigned int matrixA[XORWOW_PRECALC_MATRICES_SZ];
     unsigned int matrixB[XORWOW_PRECALC_MATRICES_SZ];
-    std::copy(matrix, matrix + XORWOW_PRECALC_MATRICES_SZ, std::begin(matrixA));
+    std::copy(matrix,
+              matrix + static_cast<std::ptrdiff_t>(XORWOW_PRECALC_MATRICES_SZ),
+              std::begin(matrixA));
     while(bool(power))
     {
         if(bool(power & 1))
@@ -132,10 +135,10 @@ inline float uniform_distribution_emu(size_t v)
     return ROCRAND_2POW32_INV + (v * ROCRAND_2POW32_INV);
 }
 
-inline void xorwow_skipahead_emu(unsigned long long skp,
-                                 prngStates* state,
-                                 const unsigned int skipahead_mat[XORWOW_PRECALC_MATRICES_NUM]
-                                                                 [XORWOW_PRECALC_MATRICES_SZ])
+inline void xorwow_skipahead_emu(
+    unsigned long long skp,
+    prngStates* state,
+    const unsigned int skipahead_mat[XORWOW_PRECALC_MATRICES_NUM][XORWOW_PRECALC_MATRICES_SZ])
 {
     unsigned int xor_vec[XORWOW_DIM];
     unsigned int* p = &(state->x);
@@ -144,10 +147,9 @@ inline void xorwow_skipahead_emu(unsigned long long skp,
     unsigned int mat_idx = 0;
     while(bool(skp)
 #if(XORWOW_PRECALC_MATRICES_NUM * XORWOW_JUMP_LOG2) < 64
-          &&
-          mat_idx < XORWOW_PRECALC_MATRICES_NUM
+          && mat_idx < XORWOW_PRECALC_MATRICES_NUM
 #endif
-          )
+    )
     {
         for(unsigned int i = 0; i < static_cast<unsigned int>(skp & XORWOW_JUMP_LOG2_MASK); i++)
         {
@@ -361,20 +363,21 @@ void DropoutBackwardVerify(const miopen::DropoutDescriptor& DropoutDesc,
                     out_len,
                     out_str);
 
-    par_ford(in_len[0], in_len[1], in_len[2], in_len[3], in_len[4])([&](
-        int i0, int i1, int i2, int i3, int i4) {
-        size_t oi =
-            out_offset + i0 * out_str[0] + i1 * out_str[1] + i2 * out_str[2] + i3 * out_str[3] + i4;
-        size_t ii =
-            in_offset + i0 * in_str[0] + i1 * in_str[1] + i2 * in_str[2] + i3 * in_str[3] + i4;
-        size_t ri = rsvsp_offset + i0 * in_len[1] * in_len[2] * in_len[3] * in_len[4] +
-                    i1 * in_len[2] * in_len[3] * in_len[4] + i2 * in_len[3] * in_len[4] +
-                    i3 * in_len[4] + i4;
+    par_ford(in_len[0], in_len[1], in_len[2], in_len[3], in_len[4])(
+        [&](int i0, int i1, int i2, int i3, int i4) {
+            size_t oi = out_offset + i0 * out_str[0] + i1 * out_str[1] + i2 * out_str[2] +
+                        i3 * out_str[3] + i4;
+            size_t ii =
+                in_offset + i0 * in_str[0] + i1 * in_str[1] + i2 * in_str[2] + i3 * in_str[3] + i4;
+            size_t ri = rsvsp_offset + i0 * in_len[1] * in_len[2] * in_len[3] * in_len[4] +
+                        i1 * in_len[2] * in_len[3] * in_len[4] + i2 * in_len[3] * in_len[4] +
+                        i3 * in_len[4] + i4;
 
-        din[ii] = static_cast<T>(bool(reservespace[ri]) && !miopen::float_equal(dropout_rate, 1.0)
-                                     ? dout[oi] / (1 - dropout_rate)
-                                     : 0);
-    });
+            din[ii] =
+                static_cast<T>(bool(reservespace[ri]) && !miopen::float_equal(dropout_rate, 1.0)
+                                   ? dout[oi] / (1 - dropout_rate)
+                                   : 0);
+        });
 }
 
 #endif
